requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai>=0.3.0
python-dotenv==1.0.0
waitress==3.0.0 